+++ App/Controller/ControllerInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
-use App\db\DbConnectInterface;
+use App\DB\DbConnectInterface;
 
 interface ControllerInterface
 {
