+++ App/DB/DbConnectInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\db;
+namespace App\DB;
 
 interface DbConnectInterface
 {
