+++ App/View/footer.phtml
@@ -1,0 +1,2 @@
+</body>
+</html>
